Update types for proper hints on lib import
