+++ src/ui/index.ts
@@ -1,6 +1,6 @@
 import { getState } from '/src/state'
 import DropZone from '/src/ui/DropZone.svelte'
-import { InkValues } from '/types/ink'
+import { InkValues } from '/types/values'
 
 import type InkInternal from '/types/internal'
 import type InkUi from '/types/ui'
