+++ types/values.ts
@@ -1,0 +1,13 @@
+export namespace InkValues {
+  export enum Appearance {
+    Dark = 'dark',
+    Light = 'light',
+  }
+
+  export enum Extensions {
+    Appearance = 'appearance',
+    Attribution = 'attribution',
+    Images = 'images',
+    Spellcheck = 'spellcheck',
+  }
+}
